ECOM-9 ECOM | Create register Product API. Closes #4.
 - Modify port to 8100.
 - Implement POST API to create product.
 - Fix issues with entity id and dto, service, controller id data type.
 - Fix mapping dates from UTC to local date formats.

+++ ms-product/src/main/java/dev/sunny/msproduct/service/ProductService.java
@@ -6,9 +6,9 @@ import java.util.Optional;
 
 public interface ProductService {
     List<ProductDto> findAllProducts();
-    Optional<ProductDto> findProductById(Integer id);
+    Optional<ProductDto> findProductById(Long id);
     ProductDto saveProduct(ProductDto productDto);
-    ProductDto updateProduct(Integer id, ProductDto productDto);
-    void deleteProduct(Integer id);
+    ProductDto updateProduct(Long id, ProductDto productDto);
+    void deleteProduct(Long id);
 }
